Add RTK_WIFI class for enhanced WiFi handling
- Introduced a new header file myWIFI.h containing the RTK_WIFI class.
- The class simplifies WiFi management with methods for configuring soft AP and station modes, handling events, and managing connections.
- Added support for ESP-NOW and various WiFi protocols.
- Included methods for setting IP addresses, SSIDs, and DHCP configurations.
- Implemented verbose debugging options and connection management features.

+++ components/ESP32_BleSerial/src/BleBufferedSerial.cpp
@@ -35,7 +35,7 @@ void BleBufferedSerial::begin(const char* name, int led_pin)
 	started = true;
 
 	char _taskName[20]; // Store random task name
-	snprintf(_taskName, sizeof(_taskName), "BLEFlushTask_%04X", esp_random() & 0xFFFF);
+	snprintf(_taskName, sizeof(_taskName), "BLEFlushTask_%04X", static_cast<unsigned int>(esp_random() & 0xFFFF));
 	xTaskCreate(flushTaskEntry, _taskName, 4096, this, 1, nullptr);
 }
 
@@ -53,7 +53,7 @@ void BleBufferedSerial::begin(const char* name, const char* service_uuid, const 
 	started = true;
 
 	char _taskName[20]; // Store random task name
-	snprintf(_taskName, sizeof(_taskName), "BLEFlushTask_%04X", esp_random() & 0xFFFF);
+	snprintf(_taskName, sizeof(_taskName), "BLEFlushTask_%04X", static_cast<unsigned int>(esp_random() & 0xFFFF));
 	xTaskCreate(flushTaskEntry, _taskName, 4096, this, 1, nullptr);
 }
